requirements.txt
fastapiuvicorn[standard]
torch
transformers
datasets
bitsandbytes
accelerate
scikit-learn